configure new structure
