+++ app/api/utilities/string.py
@@ -20,7 +20,6 @@ class StringHelper:
         string = string.strip()
         # Remove backslashes
         string = string.replace("\\", "")
-
         return string
 
     def is_url(self, string: str) -> bool:
@@ -30,3 +29,11 @@ class StringHelper:
             return all([url.scheme, url.netloc])
         except ValueError:
             return False
+
+    def slugify(self, string: str, symbol: str) -> str:
+        """Makes lowercase and strip spaces and replace space between words with "+" symbol"""
+        string = string.lower().strip()
+        string = re.sub(r"[^\w\s-]", "", string)
+        string = re.sub(r"[\s_-]+", f"{symbol}", string)
+        string = re.sub(r"^-+|-+$", "", string)
+        return string
